Updated to cache and cleaned up the spacing of the show titles

+++ rss/dvdsreleasedates-rss.sh
@@ -2,6 +2,16 @@
 
 # Requires: brew install tidy-html5
 
+CACHE=0
+while getopts ":c" opt; do
+  case ${opt} in
+    c ) CACHE=1
+      ;;
+    \? ) echo "Usage: cmd [-c]"
+      ;;
+  esac
+done
+
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 DOWNLOAD_URL="https://www.dvdsreleasedates.com"
 DOWNLOAN_FILE="dvdsreleasedates"
@@ -12,6 +22,8 @@ tidy --vertical-space no -b -q -i -asxml ${DIR}/${DOWNLOAN_FILE}.html > ${DIR}/$
 xsltproc ${DIR}/dvdsreleasedates.xslt ${DIR}/${DOWNLOAN_FILE}.tidy.html
 
 ## Cleanup
-rm ${DIR}/${DOWNLOAN_FILE}.html
-rm ${DIR}/${DOWNLOAN_FILE}.html.bak
-rm ${DIR}/${DOWNLOAN_FILE}.tidy.html
+if [ "$CACHE" -eq "0" ]; then
+  rm ${DIR}/${DOWNLOAN_FILE}.html
+  rm ${DIR}/${DOWNLOAN_FILE}.html.bak
+  rm ${DIR}/${DOWNLOAN_FILE}.tidy.html
+fi
